Assign names to the Minetown Watch.
This introduces infrastructure for randomly assigning a name from a
list to a monster, and applies it to the Minetown Watch.  So now you
can easily tell if a watchman you are seeing is the same as the one
you saw a bit ago -- if you pay attention to the names at all.

I'd like to also use this to make the petname preferences be lists
that (at least optionally) get automatically used whenever you tame
things, but that requires list preferences, so that one might have
to wait a bit.