First part of the Int/Wis split and Pw regen rebalance.
This is incomplete, but in order to test it I need to pull
in FIQ's fix for a wand-balance bug that is preventing me
from casting spells currently.  See the changelog changes
for what this commit _intends_ to do, and I'll have more
of a look at seeing if it actually works as intended once
I can test it.