Implement passive drain life attacks.
This adds the infrastructure for passive drain life attacks, as well
as a monster that has such an attack, blood puddings.  Tiles support
for blood puddings is included, as well as code that causes
sufficiently high-level puddings to create them when splitting.

Blood puddings do not generate normally, so apart from potential
bones, this should not affect most players at all.  It does add an
element of potential danger to extreme pudding farming, expecially
if you do it with long copy-paste scripts or otherwise don't pay
nearly enough attention.