Add "jump to" links to top of monster spoiler
I've added a nav bar of "Jump to" links to the top of the monster
spoiler, which jumps you to the beginning of each mlet section
in the list.

The name anchors use the index number, so they won't be stable across
versions if the monster list changes.  But the spoiler file would be
for a different version in that case anyway, so that seems ok.  And
it nicely dodges the question of whether all the mlet characters are
legal as part of an HTML name anchor -- I suspect some of them aren't.

The objects spoiler already has something similar (although the links
aren't in the objects spoiler itself, but in the spoiler index; it
will do for now).

(cherry picked from commit a89612182d5cc20fc3cb88b5192b43760eea1254)

Conflicts:
	libnethack/src/spoiler.c