Don't stop travel when walking onto an item.
From the comment, it was clearly intentional, but I don't know why --
with running, the item might be new, but who uses travel to explore? And
even if you did, it's now the same as autoexplore.

For the usual use case of travel, moving across levels that have already
been cleared, this should be a huge improvement.