Make history search actually a substring search, as the arg name implies.
The argument's being named "substring" seemed to imply that it was a
substring search, but it was in fact using strcmp, which doesn't do
that at all (and is in fact meant for sorting comparisons).  Change
the historsearch function to use strstr for the comparison, which is
actually a substring search.

Incidentally, this fixes a scoring bug that was preventing most of the
"major goals" achievements from showing up in your score when
accomplished (entering the Soko zoo, killing Vlad or the Wizard,
completing the quest, performing the invocation ritual).