Newt corpse buff:  give the buzz if it is really needed (Pw is low).
In the early game when the player's Pw and max Pw are still quite low,
make newt corpses much more likely to increase the Pw and max Pw.

Once you pass a certain threshhold, the odds revert to what they were
in vanilla.  Call it the law of diminishing returns, if you will.