Make Trollsbane light up and instantly petrify trolls.
As a side-effect of how I chose to implement this, Sunsword
will also do the same thing.  I judged this acceptable because,
although Sunsword is not as denegrated by players as Trollsbane,
it is also not widely used or highly regarded.  (The Wiki has
for years evoked snorts of derision in the community with its
description of Sunsword as "rather useful".)  The fact that
there is another weapon that can do the same thing doesn't
make Trollsbane any less useful against trolls -- nothing is
strictly _better_ against trolls than Trollsbane.