Fix bugs in the race-based skill cap augmentation.
There were two bugs here, one in the player's favor and one against.

The count wasn't being respected, so the cap was always being bumped
to the maximum.

However, the game was ignoring the maximum if it had previously
been zero.  Raising the current skill level to unskilled if it
was previously restricted fixes this.