Sokoban:  add new soko2 level, "The Snake", based on Slash'em Soko1c.
When I removed the iron bars, the rooms behind them seemed pointless.
So I removed those, but then the positioning of the pit corridor
seemed a little odd.  Most non-soko1 sokoban levels have a room with a
door on it at the end, so I added that, and then the position of the
pit corridor seemed more odd.  So I rotated the puzzle portion of the
level, and shifted the position of the ending room, and now the level
looks a bit like a snake.  Admittedly it's a bit unclear as to which
kind of snake:  some features are visually reminiscent of a cobra, and
others more like a pit viper.  However, NetHack has both of those
kinds of snakes, so whatever.