Add Sokoban level, Sneak Preview, and fix pits to holes on two others.
This new level, Sneak Preview, is based on Slash'em's Sokoban level 3f
(according to NetHackWiki numbering).  I removed the iron bars from
the location where you'd have to pass them repeatedly but kept and
even expanded the ones in the side room.  Also, the level is rotated,
condensed by one line, and adjusted so that there are spare boulders.
Finally, I added a sink, just to give the level an extra little point
of interest to distinguish it from the others better.

Also, Through the Cracks (because it used to be Soko1) and Either Way
It's All Good (due to an oversight when using soko2des.pl) previously
had pits where they should have had holes.  This is now fixed.