Replace the ring-prize Sokoban levels with Ringing Endorsement.
Ringing Endorsement is based on Slash'em Sokoban 3c, adapted to be a
prize level.  The prize here is a sack containing three items, one of
which is specified as a ring of polymorph control, one a ring of
teleport control, and one a random ring.

The logic that tries to give you Sokoban prizes you don't already have
has been changed so that having one of the control rings doesn't
always just give you two copies of the other.

The 4.3.0.1 ring-prize levels, Ring 'em Around and Jumpin' All Around
have been removed for the time being.