Seduction rebalance.
Foocubi are supposed to be dangerous, if you aren't prepared, don't
take adequate precautions, or don't know exactly what you're doing.

In vanilla, it seems that the game actively encourages players to
deliberately dally around with these demonic beings, for strategic
benefit, to the point where doing so has almost become a part of the
standard strategy among knowledgeable players.  (The overpowered
nature of unicorn horns is a significant culprit for this.)

So this is an attempt to adjust the balance.  There are still
potential benefits, but the balance is shifted so that they are
now predominantly dangerous or harmful.