Revert "Redraw the screen every hundred turns, for the benefit of dgamelaunch."
This reverts commit 63cde972e35a9440a9baeb10b6e61a18e3d6c395, because
it doesn't actually work.  Well, I mean, it does what it says -- redraws
the screen every hundred turns -- by redrawing every individual character
on the screen.  But it all happens in libnethack.

It turns out that the client has a special-case check for the redraw
command and, when the user issues that command, sends the characters
that dgamelaunch wants to see to the terminal, in addition to telling
the game to redraw the screen.  So the part of the redraw that we
actually care about happens entirely in the client, and thus doing a
periodic redraw in the actual game is unhelpful.