Add documentation for the major sets of key bindings.
This adds an SVG file (and a rendered PNG version of it) depicting
three major sets of key bindings for which we have mass rebinds:  the
traditional hjkl/yubn bindings (default), the alternative axwd/qezc
bindings, and the Dvorak-compatible uydx/pfkb bindings.

I didn't do a separate diagram to show the alternative uses for
hjklyubn when using exclusively digits or cursor-control keys for
movement, as it's only a few keys different from the defaults.

Copyright information is included.