Second attempt at redrawing the screen periodically.
In the client this time.  Unfortunately the place where it's
safe to do the redraw doesn't have access to the struct with
the moves counter from the game engine.  So I kludged it by
keeping our own counter with a static variable.  This is kind
of horrid, but casual testing suggests it may not cause any
really substantial problems.  Further testing will happen as
soon as we get this rolled out on a dgamelaunch server.