Bug fix: make race-based peacefulness work again for gnomes and dwarves.
When I split out MRACE_FOO to its own variable, I neglected to fix the
macros that calculate whether a particular monster will automatically
be peaceful or hostile purely because of your race.  The macros were
using mflags2, because that's where the race was stored before, and
with the change they now need to use mflagsr of course.

Many thanks to FIQ for tracking down the details of this one.