Adjust the formula for how many monsters Summon Nasties tries to create.
With the logic of the nasty() function now clear, the number of
monsters actually generated became a close reflection, at least in the
typical case, of the target number of monsters it was trying to make.
As soon as this was the case, testing immediately revealed that the
target number was often too low (frequently just 1 or 2).  This change
to the formula makes it always attempt to generate multiple monsters.