Rebalance starting inventory for Wizards.
I moved the marker and the spellbook of magic missile to the magic
chest contents and nuked a bunch of magic items that aren't relevant
to spell casting, which is now the One True Way to play wizards.  I
left a wand of nothing so that wizards start with basic skill in wand
use, which seems flavor-appropriate, and I left the cloak of MR,
partly to avoid overnerfing and partly because it means as soon as
they get magic missile they can immediately start using it, without
reflection, and thus it enables a wizard-appropriate playstyle.