Use a separate set of object class probabilities in the maze zone.
Vanilla has several distinct sets of probabilities for randomly
generated objects: one for objects in chests and boxes, one that
is used on the Rogue level, one that's used in Gehennom, and a
general-purpose one that's used everywhere else.

This adds an additional set of probabilities that's used in the Maze
Zone (between Medusa's level and the Castle).  Of all the sets, this
one has the most even distribution of probabilities between the
different object classes.  This makes it a reasonably good place to
encounter types of objects that aren't very likely elsewhere, just in
general; but it's not _very_ good for any particular class.  You're
as likely to find a ring as you are a weapon, for instance.