Every once in a while, an object is generated greased at random.
I've made this rare enough that it will hopefully catch players by
surprise when it does happen.  The balance implications should be
quite minor -- grease is temporary anyway, and I'm pretty sure the
frequency of objects appearing pre-greased is going to be lower
than the frequency of finding a can of grease.  But I like having
unusual things happen occasionally, for variety.