Minor adjustment to Creektown watch placement.
The original placement had one of the watchmen in the deli.
It was possible in some cases for this to result in the shop
being abandoned, if the shop creation code attempted to place
a shopkeeper on the watchman's tile.  I moved the watchman to
just outside the shop to prevent this.

Creektown also has the "design doesn't work with larger ROWNO"
problem, but that issue is not at all unique to Creektown and
will need to be addressed separately.