Vanilla bugfix: dipping a lit candle into an unlit potion of oil.
In a semi-recent rgrn thread, it was noted that dipping a lit candle
into an unlit potion of oil resulted in the generic "Interresting..."
message, and nothing else happens.  The oil is not used up.

Yeah, I know it is _possible_ to dip just the butt end of a tapered
candle into something while the other end is lit...  but that doesn't
seem very "Interresting" to me.

With this change, dipping a lit candle into an unlit potion of oil
now causes the oil to become lit, which seems like the bare minimum
that should happen, especially in NH4 where the oil is pre-IDed.