Fix big that caused two Sokoban levels to have falling rock traps.
The first level of Sokoban has to have the hardfloor flag, and pits
instead of holes, because of reasons pertaining to long-standing
limitations of the dungeon structure.  The subsequent Sokoban levels
can and should have holes, but if they are inadvertently given the
hardfloor flag, like the Soko1 levels, then the holes are instead
generated as falling rock traps, because you can't have a hole on a
level with the hardfloor flag unless it's the Castle, which is
special-cased.

Through the Cracks had the hardfloor flag because it used to be a
Soko1 level.  When I moved it up and converted its pits to holes,
I neglected to remove the hardfloor flag.

Either Way, It's All Good had the hardfloor flag by mistake as a
consequence of the manner in which the level was made.  (I used
hackedhead's nhss sokoban solver for testing, then used a Perl
script, libnethack/utils/soko2des.pl, to get it into .des file
format.  At the time, soko2des.pl always included the hardfloor
flag, because when I first wrote it I was doing Soko1 levels.)