Fix bugs in the dragon-hall filling code.
The major one here is an off-by-one that could cause a dragon to be
placed at a position dependent on unitialized memory.  In the common
case, the dragon would usually end up at 0,0, which is weird if you
have monster detection and otherwise mostly harmless, but in exotic
cases this could be an exploitable bug).

I also cleaned up an unused-variable warning.