Support the ability to receive dgamelaunch simple mail messages.
At this time, only the ability to receive mail is implemented, so only
dgamelaunch-based servers are supported.  It would be nice to also
have the ability to create messages when watching another player live,
but I haven't implemented that yet.

Unfortunately, this feature breaks save compatibility (because the
option needs to be saved, otherwise it would turn itself off every
turn and nobody will ever actually receive any messages), so it will
not go into 4.3.0.3.  Oh, well.