Magic cancellation rebalance for monster-vs-monster combat.
I doubt most players will notice this, but the mhitm cases now handle
magic cancellation/negation in, to a large extent, roughly the same
manner as the monster-vs-player case.  There are a number of specifics
that don't entirely match, because things aren't quite the same for
monsters as players, but I tried to make the effect of MC comparable
wherever possible.

I also cleaned up one case in the monster-vs-player rebalance that
potentially could've caused a divide-by-zero situation; now it won't.