Advent calendar enhancements: colored lights, candy canes, engravings
This puts the door numbers in colored lights, adds engravings behind
the doors explaining the advent (as is traditional for advent
calendars), replaces the fortune cookies with candy canes and sugar
plums (via fruitname), and makes balance adjustments to the prize
behind the 24th door.  Challenge mode prevents you from getting the
very best possible gifts, replacing them with some not-so-OP ones.