Make shopkeepers consistent about the prices they offer you.
When you drop an item in the shop, sometimes the shopkeeper offers you
the maximum price he would ever be willing to pay, and sometimes he
only offers you 3/4 of that price.

In 3.4.3, this decision is made using a random number, so you can
get both prices by repeatedly dropping the item.  This makes price
checking more tedious, because the optimal thing to do, and the
standard technique therefore among experienced players, is to drop
each item repeatedly (typically, half a dozen times or so) until
you get a second price.  Then you take the larger price, double it,
and have the "base value" of the item.

This makes the 25% reduction in what the shopkeeper is willing to
pay consistent for any given type of item, so there's no longer any
point in dropping one repeatedly.