Implement missing elemental melee attacks for the new dragons, monster-vs-player
This implements the mhitu case for (non-passive) melee attacks that do
magic or disintegration damage, for gray and black (respectively)
elder and great dragons.

This would make great black dragons, with their 16 speed, rather scary
to a non-fast player.  One such dragon can potentially wipe out your
cloak and body armor in a single turn, and you are not fast enough to
flee, and reflection is useless against melee attacks.  However, the
black versions of elder and great dragons are currently not generated
(there are no black dragon halls, and elder and great dragons of any
color are not generated normally, for balance reasons.

I may in the future find excuses to put limited numbers of these
dangerous beasties in limited situations, though.