Redesigned Monastery of the Earth Lord (Monk quest locate level)
The old one was basically just a cave with a room in the corner, and
it was also a duplicate of one of the Rogue quest levels, plus also it
was on ais523's list of levels that need to change because they are 21
rows tall and thus prevent him from reducing ROWNO by one (to allow
an extra line for something else in 80x24 mode, which he still wants
to support in NH4).

So I redesigned it.