Redesigned Assassin's Guild Hall level (Rogue Quest Locate)
The Rogue quest locate level, in addition to being on ais523's list of
21-row levels, was also a duplicate and not really thematically
appropriate for its flavor text.  There was nothing really about it to
indicate that it was an Assassin's Guild Hall at all.  So I've
redesigned the level.  Now as you cross it monsters will try to
assassinate you with ranged attacks from behind iron bars.  To
compensate for this added danger, I have somewhat increased the
level's quota of loot.