Fully hook up shield skill to player AC.
Instead of just improving your chance to dodge, wearing a shield at
basic shield skill or better now actually improves your Def.  At high
skill levels, this is multiplicative with enchantment (unless the
enchantment is worse than +1, in which case it is additive, so that
more skill is still better than less skill).