Advent calendar:  don't switch in/out positions when ROWNO is large.
The original patch assumed normal values of ROWNO and so compared y to
a constant (10) when deciding whether a given door was in the top half
or bottom half of the level.  This failed when compiled with a large
value of ROWNO, resulting in the numbers being inside the locked
closets and the candy canes and sugar plums and story elements being
on the outside, in front of the doors.  By comparing to (ROWNO / 2),
we can get the correct layout regardless of how large ROWNO is, since
the geometry line in the .des file centers the level vertically.