Clarify in README-MSWindows.txt that GetGnuWin doesn't install flex or bison.
As near as I can tell, GetGnuWin does not have the capability to
install any GnuWin software except GetGnuWin.  If it does, the
capability is so well hidden as to be useless.  So I advise the
user to download the flex and bison packages, which do work.

I also clarified the SDL situation:  it isn't needed unless you
want graphical tiles, IMO.