Fourk-specific DEADMONSTER() changes.
Clean up a few cases where a monster's mhp was being checked in code
that isn't in NH4 master.  Some of this may be 4.3-maint stuff, some
of it is L patch, and some of it may be my own code, not sure.

Hopefully I got all the relevant cases.  I think most of the remaining
comparisons between mhp and 0 or 1 are correct on the grounds that they
happen when mhp has changed and are done to decide whether to call one
of the various kill-the-monster functions (xkilled, mondead, etc.)