Use trap background color for stinking cloud areas, in the text tilesets.
The ascii_overrides are used by the ASCII and Unicode tilesets, so
both of them will have this change now.  Stinking cloud areas will
show up with a cyan background.

This is arguably a gameplay change, as it lets the player see into
stinking-cloud areas.  (However, being in a stinking cloud still
causes blindness if you are not immune.)