Update challenge mode documentation.
Update the challenge mode documentation so that it documents
challenge-mode differences in recent enhancements, including
the advent calendar.

Also, ensure that only integer arithmetic is used when calculating
the higher max monster difficulty in challenge mode.  There's no
telling when someone will be using an original Pentium or something,
so floating point numbers are not to be trusted.