Conduct tracking for conflict, invisibility, and displacmenet.
If you complete the game without using conflict, without ever being
invisible, or without ever being displaced, you game dump messages and
the xlogfile entry will now reflect this fact.

If you break the conducts, there's no message, because that would get
spammy.  Discoverability is preserved because you can see the messages
in the event of early death.

The in-game conduct list will also show that you haven't been
displaced, if that is the case, because it's impossible to be
displaced and not know it.  However, due to unidentified rings in the
case of conflict and unidentified potions and blindness in the case of
invisibility, it's possible to break these conducts without knowing,
so the messages for them are only shown at game end for now.