Issue a message when a stinking cloud trap is triggered.
I was relying on the function that creates the cloud to print a
message if the player can detect the cloud.  It does not, because it
was originally designed to be called when reading a scroll, which
tells you that it's a scroll of stinking cloud and asks you where to
center the cloud, making such a message redundant.  So the caller must
print any message that's wanted in other cases as well.