Take player race into account (when not polyselfed) in more situations.
A lot of the existing NetHack code is written to assume that a player
who is not polymorphed is ipso facto human or very similar to human.

Among other things, a player who is not polyselfed can always wear all
the same armor as a human.  This is now fixed, with the implication
that scurriers and gnomes, being small, cannot wear body armor.

A previous commit had already fixed pickless digging, so that scurriers
can now use their racial digging ability.

Scurriers are promoted to small (formerly they were tiny) so that they
can wear other armor.  In particular, they can still wear boots, ergo
can still get extrinsic speed and thus be faster than other races.

The following additional actions are _currently_ unchanged, because
there is no playable race for which they are relevant; but your race
is now taken into account (in your natural form) for determining your
ability to do the following things:
 * blow an instrument (such as a whistle) that requires blowing.
 * wielding a weapon -- your race must be capable of wielding.
 * attacks that stick you to the opponent.
 * intrinsic kicking ability.
 * emitting light
 * riding
 * hiding under objects (this may be only partially supported)
 * #monster
 * horns block hard helmets.
 * having hands and is required to put on any armor, even if it fits.

All of this is coded in a way that supports sexual dimorphism automatically,
so for instance if we add a "firefly" race, it would be possible for only
the males of that race to emit light and not the females.  (I have no plans
to make firefly a monster, let alone a playable race.  It's just an example.)

This is just a start.  There are many additional situations that have
not yet been addressed.