Quasits and invisibility
Quasits were still kind of boring, so I looked them up on the
internet, and it turns out that in the source material, and in a
number of other roguelike games (notably the Angband family) they are
typically invisible.  So I've made quasits invisible a percentage of
the time (25% in normal mode, 75% in challenge mode).

I also gave them see invisible (always), because it's weird to have
invisible creaters that can't see invisible, unless there's a specific
flavor reason for that, which is lacking here.