Don't blame player for displacing pet into an undetected trap.
Getting blamed for displacing a pet into a known trap is one thing.
But if you don't know the trap is there, it seems just a bit unfair.
This is particularly relevant for early-game Valkyries, who due to
flying may not have a reliable way to detect certain kinds of traps
even after moving onto them.

Thanks to hackedhead for noticing this situation.