Use os_invalid rather than os_tool as a sentinel value.
The Fourk sacrifice gift code has a variable for "this is the slot
we're trying to generate something to fill"; but sometimes it hasn't
decided on a particular slot and so needs a sentinel value for that.
Previously, it was using os_tool, but os_invalid is a better choice.
Thanks to FIQ for pointing this out.