Teach body_part how to handle player race.
For example, if you are a scurrier and have dropped your weapon for
some reason, the message is now "You begin bashing monsters with your
bare paws" (or "gloved" if that is the case).  Previously the message
said "hands" regardless of your race, because your role was considered
human.  This was mostly not a big deal for races like elf and gnome,
but it doesn't work as well if we're going to have races that are a
bit less like humans than that.