Minor Sokoban fix, Either Way, It's All Good: wand and ring locations
Prevent the guaranteed wand and ring in this version of Sokoban from
being generated in a location that might become inaccessible before
the player reaches it.

Because of the two-way nature of the level, the player upon exiting
the last pit corridor may trap a boulder in a location that he has
had no previous opportunity to examine.  This change prevents either
the wand or the ring from being generated there.