Make Mine's End consistent about whether it has a magic chest or not.
The new version, Orc Temple, was lacking a magic chest.  I added it.

There remain questions among some players as to whether Mine's End
should have a magic chest and, indeed, whether there are too many of
them in the game at present, or too few, etc.  This commit does not
attempt to address all of that.  I just made the Orc Temple level
consistent with the other Minetown levels in this regard, for now.