Dragonbane enhancements, round one.
Dragonbane is now a lance, and knights receive it when crowned
(provided it does not already exist).

Also, FWIW, lances now get the kebab bonus, same as spears.

Note that, although it's impossible to test this in 3.4.3, an
investigation of the jousting code reveals that, if I read the
conditionals correctly, artifact lances shouldn't break when jousting.
(If Dragonbane does break when jousting, that'd be a bug.  I intend
that it should not.  I was going to change the joust code to ensure
this, but I now believe no change is necessary.  TDDTOE.)

Thanks to hackedhead and FIQ for helping me think through what to do
with this artifact to rescue it from the "completely useless" pile.
It may need further enhancements, and there's still one idea on the
table (a suggestion that dragons always drop scales if killed with
it), but this is a start.