Explicit tracking for when monsters are tracked in an iceblock.
I was getting reports from players that polymorph vs non-player
monsters was creating stone at times when it should not, when there
was no evidence of ice devils on the level.

If there were in fact no ice devils when it was happening, this kind
of implies that somehow monsters were ending up with mtrapped set when
it should not have been, when there were no traps on their tile.  I
suspect the code could use an audit for that, as it might cause other,
subtler problems, unless I'm missing something.

But in any case, the encased-in-iceblock situation for monsters is now
tracked explictly, so the known visible symptoms should be cleared up.