Guarantee a couple of xorns at the Castle.
In 3.4.3, the random monsters on the Castle level almost always
include xorns, because they are the only member of the X mlet, which
is on the Castle's list of random monster letters.  In Fourk, there
are several other X monsters, so the random ones are much less likely
to be xorns.

To compensate for this, guarantee a couple of xorns when the level
is first generated.