Quarterstaff rework, round one.
 * Quarterstaves do less damage in melee than before.

 * Wielding a (non-negatively-enchanted) quarterstaff now
   helps partially compensate for worn armor, when determining
   spell failure rates.  This is less effective than (but stacks
   with) wearing a robe and can, for some spells, bring your
   failure rate to 0% when it otherwise would not be.

 * Additionally, wielding a (non-negatively-enchanted) quarterstaff
   improves the base chance of casting spells that are just a bit
   above your level, or when your Int is just a bit low, etc.
   Enchanting the quarterstaff improves this effect.

Thanks to Mandevil and FIQ for feedback at the idea stage on this.

This is only the first round of changes for quarterstaves.  I expect
that these changes will need to be reviewed after some play testing.
The numbers will probably want to be tweaked somewhat, but I'm not yet
sure in which direction.