Show prayer in the history.
Prayer is now a non-hidden non-livelogged historic event, which the
player can see in-game in the list of historic events.

Thanks to hackedhead for pointing out that what turn you last prayed
wasn't shown anywhere in the interface.

Whether it's safe to pray or not isn't shown (usually, when you don't
have enlightenment) because the character doesn't know this; but when
you last did pray is something the player theoretically knows.  Going
through the replay interface to retrieve this information is rather
inconvenient, and requiring the player to keep notes external to the
game itself goes against NH4 philosophy and the "InterHack argument".

There was some talk about putting this information under conducts, but
to me the game history seemed a more logical place for it.