Wizard quest locate level, redesign.
As part of the project to limit special levels to 20 rows, the Wizard
quest locate level needed to have a row trimmed off.  That would have
been difficult to do with the existing level design.  The existing
level design was also not particularly interesting, so rather than
try to make difficult changes to a not-so-great level design, I just
designed a replacement level from scratch, around the level's title
("Tower of Darkness").

While I was doing the redesign anyway, I went ahead and built in a
couple of desirable spellbooks, so that wizards who gain access to
the quest can obtain them.  This is balanced by the fact that they
are locked in small rooms with somewhat challenging monsters.