Cloak of Half Physical Damage
Add a new random-appearance magic cloak, the cloak of half physical
damage.  For balance reasons, this cloak does not provide MC.  It
also weighs more than the other cloaks.

The new random cloak appearance is "dirty rag", which I chose because
it was already a random cloak appearance in Slash'em, so some tilesets
already have it.