Make xlogfile and livelog field separator customizable, default to tab.
Previously, the field separator in the xlogfile and livelog files was
always colon.  Now it defaults to a horizontal tab.  Just in case
anyone is not pleased with this change, you can simply add
-DXLOGFILE_FIELD_SEPARATOR=":" to your CFLAGS environment variable,
or in some other way define XLOGFILE_FIELD_SEPARATOR at compile time,
and all will be as it was.

In the unlikely event that anyone wants the livelog field separator to
be different from the xlogfile one, define LIVELOG_FIELD_SEPARATOR and
voila, Bob is your uncle.  (Otherwise, the livelog will use the same
field separator as the xlogfile uses.)