Fix Minetown Watch inventory bug (memory leak)
The inventory items for the Minetown Watch were being generated, but
they were being added to the original struct monst, after the
christening function had created a new, named struct monst.  Fix it.

Thanks to Khor for noticing this amazing bug, and thanks to FIQ for
pointing out that christening a monster causes reallocation, a fact I
had failed to notice.