Wishing:  cross-aligned artifact wishes are more likely to fail now.
When wishing for a cross-aligned artifact, the odds of failure are
now calculated as if there was one additional (in normal mode) or
were two additional (in challenge mode) artifacts already existing,
compared to the odds when wishing for a co-aligned artifact.

Thus, in order to be guaranteed a cross-aligned artifact from a wish,
in normal mode, there must be no more than one artifact already
existing in the game for a normal wish, or at most two artifacts
already existing in the case of a blessed scroll of wishing.

In challenge mode, to be guaranteed a cross-aligned artifact from a
wish, there must be no artifacts already existing and also you must be
using a blessed scroll of wishing.