Elbereth must be the only thing on the tile to be scary.
The name of Elbereth is now only considered "scary" when it is the
only thing engraved on the tile (although the longer form "Elbereth
Gilthoniel" also works, as does "A Elbereth Gilthoniel", which is
actually a song title).

This change was inspired by one of the changes in 3.6.0, but goes
a step further.  In 3.6.0, "F?hcr?tnElbereth" doesn't work, but
"ElberethWasHere" does.  This seemed asymetrical to me, so neither
of those engravings is considered scary in Fourk.