Add new Medusa variant from 3.6.0
This adds two new level variants to medusa.des, but only one of them
(the one with the ravens) is enabled for NetHack Fourk.  The other
(with the snakes) is not enabled, because it triggers a bug in the
Fourk codebase.  (I suspect this is likely because I haven't applied
Pasi's level compiler changes; 3.6.0 has done so.)  The bug causes the
Queen of the Snakes variant of Medusa's Island to have no rooms
defined in its structure, which causes a problem in fixup_special.

For now, I have listed the medusa level in dungeon.def as having
three variants, rather than four, so the one with the problem is
never used.

Cherry-picked from commit 5cf86e0409, from the Dev Team's public 3.6
repository.