Bugfix:  incorrect array subscript in sacrifice_gift()
When iterating over the spells, I forgot to subtract the first one
from the number used as an index into an array.  Both FIQ and
travis.ci were getting "array subscript above bounds" errors for this.
Not sure why my compiler here was not complaining, but the problem was
obviously very real once I looked at it.