Make the invocation more robust, in case the VS is placed wrongly.
If the invocation is performed near the edge of the map, trigger
an impossible() for each would-be-changed off-the-map tile, but
don't panic() unless the stairs themselves cannot be placed.

This SHOULD never be necessary, because the code that sets the
invocation position tries to keep it in bounds.  Nonetheless, a
tournament player reported an invocation position that resulted
in the panic(), so something is wrong there.  This fix should
at least allow the game to be continued (using P to get past the
impossible() dialogs).

In order to test that it works as designed, I also had to fix the
trap wizwishing code.  Wishing for the vibrating square created
a vibrating square trap, which farlook reported as such; but you
didn't feel the strange vibration and couldn't perform the ritual,
because gamestate.inv_pos was not updated.  This fixes that, and
removes the original vibrating square trap (if it's on the level)
into the bargain.

Tested, appears to work.

Additionally, I made both mkaisvs and mkinvpos slightly more
paranoid about the level's up stairs.  Again, this shouldn't be
necessary, because the retry loop should prevent such a location.
But a little extra caution may be warranted for something as
critically important as the vibrating square.