Set the Big Stick #invoke effect to conflict.
When we made the Sceptre of Might the Priest quest artifact, we
gave it the power boost #invoke effect appropriate to that role.
This left the conflict effect free.  Attaching it to Big Stick
gives it back to the role that had it in the first place.

Thanks to FIQ for suggesting this.